si - shave a couple of bytes off the transaction table rowKey
--HG--
extra : rebase_source : b4fa4ad13fc2cdf4246afda354b132230f091099

+++ structured_si/src/main/java/com/splicemachine/si/impl/TransactionStore.java
@@ -494,7 +494,7 @@ public class TransactionStore<Data, Result, KeyValue, Put, Delete, Get, Scan, Op
      * @return
      */
     private Object[] transactionIdToRowKey(long id) {
-        return new Object[] {(short) (id % 16), id};
+        return new Object[] {(byte) (id % 16), id};
     }
 
     /**
